Initial Commit. Added basic lexing functionality
Added a class ''Token'' to represent a token in the lexing stream.
It basically aggregates the type of the token, the position in the
source and a value, if it's a number or a identifier/keyword

Added a class ''Stream'' to have an iterator over the source which keeps
track of the position in the file

Added a generator function ''lex'' that reads from an ''Stream'' and
yields the token it recognizes.